Use Rust version of lodepng.
The executable as a whole gets 21% larger. This might be because we are
including Rust's libstd (it has stuff like the allocator and panic handler that
probably can't be optimized away), or because the C code is compiled with -Os
and for Rust I'm using the debug build. In any case I'm not worried about it--
data.zip is 13x the size of the executable anyway.